+++ mgmap/src/main/java/mg/mgmap/activity/mgmap/features/routing/profile/CostCalculatorSimple.java
@@ -15,7 +15,7 @@ public class CostCalculatorSimple implements IfCostCalcHeuristic {
     }
 
     @Override
-    public double calcCosts(double dist, double vertDist) {
+    public double calcCosts(double dist, float vertDist) {
         if (vertDist > 0){
             return dist + vertDist*mVertDistCostFactor + 0.0001;
         }
